Innhenting av data tar no veldig mykje plass i Brevoppretter-klassa, så skil det heller ut i ei eiga klasse med det som einaste ansvar
