+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/Brevoppretter.kt
@@ -1,34 +1,21 @@
 package no.nav.etterlatte.brev
 
-import kotlinx.coroutines.async
-import kotlinx.coroutines.coroutineScope
 import no.nav.etterlatte.brev.adresse.AdresseService
 import no.nav.etterlatte.brev.behandling.PersonerISak
-import no.nav.etterlatte.brev.behandling.erForeldreloes
-import no.nav.etterlatte.brev.behandling.loependeIPesys
-import no.nav.etterlatte.brev.behandling.opprettAvsenderRequest
-import no.nav.etterlatte.brev.behandling.somTittel
-import no.nav.etterlatte.brev.brevbaker.BrevbakerRequest
-import no.nav.etterlatte.brev.brevbaker.BrevbakerService
 import no.nav.etterlatte.brev.db.BrevRepository
-import no.nav.etterlatte.brev.hentinformasjon.BrevdataFacade
 import no.nav.etterlatte.brev.hentinformasjon.behandling.BehandlingService
 import no.nav.etterlatte.brev.model.Adresse
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevDataRedigerbar
-import no.nav.etterlatte.brev.model.BrevInnhold
-import no.nav.etterlatte.brev.model.BrevInnholdVedlegg
 import no.nav.etterlatte.brev.model.BrevProsessType
 import no.nav.etterlatte.brev.model.BrevkodeRequest
 import no.nav.etterlatte.brev.model.Mottaker
 import no.nav.etterlatte.brev.model.OpprettNyttBrev
-import no.nav.etterlatte.brev.model.Spraak
 import no.nav.etterlatte.libs.common.behandling.SakType
 import no.nav.etterlatte.libs.common.feilhaandtering.UgyldigForespoerselException
 import no.nav.etterlatte.libs.common.person.Folkeregisteridentifikator
 import no.nav.etterlatte.libs.common.person.UkjentVergemaal
 import no.nav.etterlatte.libs.common.person.Vergemaal
-import no.nav.etterlatte.libs.common.retryOgPakkUt
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
 import no.nav.etterlatte.libs.ktor.token.BrukerTokenInfo
 import no.nav.etterlatte.libs.ktor.token.Saksbehandler
@@ -37,10 +24,8 @@ import java.util.UUID
 class Brevoppretter(
     private val adresseService: AdresseService,
     private val db: BrevRepository,
-    private val brevdataFacade: BrevdataFacade,
     private val behandlingService: BehandlingService,
-    private val brevbaker: BrevbakerService,
-    private val redigerbartVedleggHenter: RedigerbartVedleggHenter,
+    private val innholdTilRedigerbartBrevHenter: InnholdTilRedigerbartBrevHenter,
 ) {
     suspend fun opprettVedtaksbrev(
         sakId: Long,
@@ -79,7 +64,7 @@ class Brevoppretter(
         brevDataMapping: suspend (BrevDataRedigerbarRequest) -> BrevDataRedigerbar,
     ): Pair<Brev, String> =
         with(
-            hentInnData(
+            innholdTilRedigerbartBrevHenter.hentInnData(
                 sakId,
                 behandlingId,
                 bruker,
@@ -113,7 +98,7 @@ class Brevoppretter(
         val spraak = db.hentBrevInnhold(brevId)?.spraak
 
         with(
-            hentInnData(
+            innholdTilRedigerbartBrevHenter.hentInnData(
                 sakId,
                 behandlingId,
                 bruker,
@@ -133,105 +118,6 @@ class Brevoppretter(
             return BrevService.BrevPayload(
                 innhold.payload ?: db.hentBrevPayload(brevId),
                 innholdVedlegg ?: db.hentBrevPayloadVedlegg(brevId),
-            )
-        }
-    }
-
-    private suspend fun hentInnData(
-        sakId: Long,
-        behandlingId: UUID?,
-        bruker: BrukerTokenInfo,
-        brevKodeMapping: (b: BrevkodeRequest) -> EtterlatteBrevKode,
-        brevDataMapping: suspend (BrevDataRedigerbarRequest) -> BrevDataRedigerbar,
-        overstyrSpraak: Spraak? = null,
-    ): OpprettBrevRequest {
-        val generellBrevData =
-            retryOgPakkUt { brevdataFacade.hentGenerellBrevData(sakId, behandlingId, overstyrSpraak, bruker) }
-
-        val brevkodeRequest =
-            BrevkodeRequest(
-                loependeIPesys(
-                    systemkilde = generellBrevData.systemkilde,
-                    behandlingId = generellBrevData.behandlingId,
-                    revurderingsaarsak = generellBrevData.revurderingsaarsak,
-                ),
-                erForeldreloes(generellBrevData.personerISak.soeker, generellBrevData.personerISak.avdoede),
-                generellBrevData.sak.sakType,
-                generellBrevData.forenkletVedtak?.type,
-            )
-
-        val kode = brevKodeMapping(brevkodeRequest)
-        val tittel =
-            kode.tittel ?: (
-                generellBrevData.forenkletVedtak
-                    ?.type
-                    ?.somTittel()
-                    ?.let { "Vedtak om $it" } ?: "Tittel mangler"
-            )
-        return coroutineScope {
-            val brevDataRedigerbarRequest =
-                BrevDataRedigerbarRequest(
-                    brukerTokenInfo = bruker,
-                    soekerOgEventuellVerge = generellBrevData.personerISak.soekerOgEventuellVerge(),
-                    sakType = generellBrevData.sak.sakType,
-                    forenkletVedtak = generellBrevData.forenkletVedtak,
-                    utlandstilknytningType = generellBrevData.utlandstilknytning?.type,
-                    revurderingsaarsak = generellBrevData.revurderingsaarsak,
-                    behandlingId = behandlingId,
-                    erForeldreloes = erForeldreloes(generellBrevData.personerISak.soeker, generellBrevData.personerISak.avdoede),
-                    loependeIPesys =
-                        loependeIPesys(
-                            generellBrevData.systemkilde,
-                            generellBrevData.behandlingId,
-                            generellBrevData.revurderingsaarsak,
-                        ),
-                    systemkilde = generellBrevData.systemkilde,
-                    avdoede = generellBrevData.personerISak.avdoede,
-                )
-            val brevData: BrevDataRedigerbar = brevDataMapping(brevDataRedigerbarRequest)
-
-            val innhold =
-                async {
-                    brevbaker.hentRedigerbarTekstFraBrevbakeren(
-                        BrevbakerRequest.fra(
-                            brevKode = kode,
-                            brevData = brevData,
-                            avsender =
-                                adresseService.hentAvsender(
-                                    opprettAvsenderRequest(bruker, generellBrevData.forenkletVedtak, generellBrevData.sak.enhet),
-                                ),
-                            soekerOgEventuellVerge = generellBrevData.personerISak.soekerOgEventuellVerge(),
-                            sakId = sakId,
-                            spraak = generellBrevData.spraak,
-                            sakType = generellBrevData.sak.sakType,
-                        ),
-                    )
-                }
-
-            val innholdVedlegg =
-                async {
-                    redigerbartVedleggHenter.hentInitiellPayloadVedlegg(
-                        bruker,
-                        kode.brevtype,
-                        generellBrevData.sak.sakType,
-                        generellBrevData.forenkletVedtak?.type,
-                        generellBrevData.behandlingId,
-                        generellBrevData.revurderingsaarsak,
-                        generellBrevData.personerISak.soekerOgEventuellVerge(),
-                        generellBrevData.sak.id,
-                        generellBrevData.forenkletVedtak,
-                        generellBrevData.sak.enhet,
-                        generellBrevData.spraak,
-                    )
-                }
-
-            OpprettBrevRequest(
-                soekerFnr = generellBrevData.personerISak.soeker.fnr.value,
-                sakType = generellBrevData.sak.sakType,
-                enhet = generellBrevData.sak.enhet,
-                personerISak = generellBrevData.personerISak,
-                innhold = BrevInnhold(tittel, generellBrevData.spraak, innhold.await()),
-                innholdVedlegg = innholdVedlegg.await(),
             )
         }
     }
@@ -265,15 +151,6 @@ class Brevoppretter(
             adresse = Adresse(adresseType = "", landkode = "", land = ""),
         )
 }
-
-private data class OpprettBrevRequest(
-    val soekerFnr: String,
-    val sakType: SakType,
-    val enhet: String,
-    val personerISak: PersonerISak,
-    val innhold: BrevInnhold,
-    val innholdVedlegg: List<BrevInnholdVedlegg>?,
-)
 
 class KanIkkeOppretteVedtaksbrev(
     behandlingId: UUID,
